+++ src/index.js
@@ -7,6 +7,8 @@ import Text from './text';
 import AMStorage from './storage';
 import AppBootstrap from './appBootstrap';
 import ObjHandler from './objHandler';
+import HttpErrorHandler from './httpErrorHandler';
+import Notification from './notification';
 
 export {
     Http,
@@ -17,5 +19,7 @@ export {
     VersionCheck,
     AMStorage,
     AppBootstrap,
-    ObjHandler
+    ObjHandler,
+    HttpErrorHandler,
+    Notification
 };