+++ CHANGELOG.md
@@ -6,6 +6,21 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.1.4/1.1.7] - 2018-02-26
+### Added
+- Client notification for cordova, native and webapp
+
+## [1.1.0/1.1.3] - 2018-02-16
+### Added
+- Error handler in Http: now the fetch errors are handled correctly
+### Changed
+- AppBootstrap now use the new Http.setup and set the callbacks for Http errors
+- Deprecated warning on Http.setHeaders
+
+## [1.0.25] - 2018-02-15
+### Changed
+- Http.reset and Http.setup created to simplify the bootstrap
+
 ## [1.0.24] - 2018-01-22
 ### Changed
 - Text.formatNumber and Text.formatReal added